pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "pixel-knitter"
dynamic = ["version"]
description = 'takes a bitmap and turns it into a knitting pattern'
readme = "README.md"
requires-python = ">=3.10"
license = "WTFPL"
keywords = []
authors = [
  { name = "Étienne Noss", email = "etienne.noss+pypi@gmail.com" },
]
classifiers = [
  "Development Status :: 4 - Beta",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: Implementation :: CPython",
  "Programming Language :: Python :: Implementation :: PyPy",
]
dependencies = ["pillow>=10.2.0", "jinja2>=3.1.3"]

[project.urls]
Documentation = "https://github.com/etene/pixel-knitter#readme"
Issues = "https://github.com/etene/pixel-knitter/issues"
Source = "https://github.com/etene/pixel-knitter"


[project.scripts]
pixel-knitter = "pixel_knitter.__main__:main"

[tool.hatch.version]
path = "src/pixel_knitter/__about__.py"

[tool.hatch.envs.default]
dependencies = [
  "coverage[toml]>=6.5",
  "pytest==8.0.0",
  "lxml>=5.0.0",
]
[tool.hatch.envs.default.scripts]
test = "pytest {args:tests}"
test-cov = "coverage run -m pytest {args:tests}"
cov-report = [
  "- coverage combine",
  "coverage report",
]
cov = [
  "test-cov",
  "cov-report",
]
lint = "ruff src/ tests/"

[[tool.hatch.envs.all.matrix]]
python = ["3.10", "3.11", "3.12"]

[tool.hatch.envs.types]
dependencies = [
  "mypy>=1.0.0",
]
[tool.hatch.envs.types.scripts]
check = "mypy --install-types --non-interactive {args:src/pixel_knitter tests}"

[tool.coverage.run]
source_pkgs = ["pixel_knitter", "tests"]
branch = true
parallel = true
omit = [
  "src/pixel_knitter/__about__.py",
]

[tool.coverage.paths]
pixel_knitter = ["src/pixel_knitter"]
tests = ["tests"]

[tool.coverage.report]
exclude_lines = [
  "no cov",
  "if __name__ == .__main__.:",
  "if TYPE_CHECKING:",
]
